Split Electron app functions in to files
